20240104 @Mookse - modularize all avatar functions - add timeouts
