+++ .env-sample
@@ -1,26 +1,27 @@
-OPENAI_API_KEY={ your individual openai API key -- see instructions}
+OPENAI_API_CHAT_TIMEOUT=22000
+OPENAI_API_CHAT_RESPONSE_PING_INTERVAL=890 # how often to ping openai service for assistant run status complete
+OPENAI_API_KEY=sk-... # your individual openai API key
 OPENAI_BASE_URL=https://api.openai.com/v1
 OPENAI_MAHT_GITHUB=https://github.com/MyLife-Services/mylife-maht.git
 OPENAI_MODEL_CORE="gpt-3.5-turbo"
-OPENAI_ORG_KEY={ use ours or use your own }
-OPENAI_TIMEOUT={ timeout }
-PORT={ KOA port }
+OPENAI_ORG_KEY=org-... # use ours or use your own
+PORT=3000 # KOA port
 MYLIFE_ALLOW_INTELLIGENT_QUESTIONS=false
 MYLIFE_DB_ALLOW_SAVE=true
-MYLIFE_DB_ENDPOINT={ get from administrator }
-MYLIFE_DB_NAME={ get from administrator }
-MYLIFE_DB_RW={ get from administrator }
-MYLIFE_DB_RX={ get from administrator }
-MYLIFE_DB_CONTAINER_NAME={ get from administrator }
-MYLIFE_SERVER_MBR_ID={ get Maht info from administrator or request own }
-MYLIFE_HOSTED_MBR_ID=["mbr_id's hosted by this server"]
-MYLIFE_SESSION_KEY={ type in own personal key, quoted or unquoted if no spaces... Restarting server will NOT refresh sessions (as of now) so must change this key to refresh server-level constants}
-MYLIFE_SESSION_TIMEOUT_MS={ session timeout in milliseconds }
-MYLIFE_VERSION=1.0
+MYLIFE_DB_ENDPOINT= # get from administrator
+MYLIFE_DB_NAME= # get from administrator
+MYLIFE_DB_RW= # get from administrator
+MYLIFE_DB_RX= # get from administrator
+MYLIFE_DB_CONTAINER_NAME= # get from administrator
+MYLIFE_SERVER_MBR_ID= # get from administrator
+MYLIFE_HOSTED_MBR_ID=[] # ["mbr_id's hosted by this server"]
+MYLIFE_SESSION_KEY=iteration-number # type in own personal key, quoted or unquoted if no spaces... Restarting server will NOT refresh sessions (as of now) so must change this key to refresh server-level constants}
+MYLIFE_SESSION_TIMEOUT_MS=900000 # session timeout in milliseconds
+MYLIFE_VERSION=0.0
 MYLIFE_EMBEDDING_SERVER_URL=http://mylife-embedding-server
 MYLIFE_EMBEDDING_SERVER_PORT=8000
 MYLIFE_EMBEDDING_SERVER_BEARER_TOKEN={ chosen personally to match environment variable in embedding server }
-MYLIFE_EMBEDDING_SERVER_FILESIZE_LIMIT={ filesize } # 1MB = 1024 * 1024
-MYLIFE_EMBEDDING_SERVER_FILESIZE_LIMIT_ADMIN={ filesize } # 10MB = 1024 * 1024 * 10
-MYLIFE_CONTRIBUTIONS_DB_CONTAINER_NAME={ get from administrator }
-MYLIFE_REGISTRATION_DB_CONTAINER_NAME={ get from administrator }
+MYLIFE_EMBEDDING_SERVER_FILESIZE_LIMIT=1024 # 1MB = 1024 * 1024
+MYLIFE_EMBEDDING_SERVER_FILESIZE_LIMIT_ADMIN=1024 # 10MB = 1024 * 1024 * 10
+MYLIFE_CONTRIBUTIONS_DB_CONTAINER_NAME= # get from administrator
+MYLIFE_REGISTRATION_DB_CONTAINER_NAME= # get from administrator
