+++ inc/js/factory-class-extenders/class-extenders.mjs
@@ -1,12 +1,20 @@
 import { EventEmitter } from 'events'
-import { Marked } from 'marked'
 import { EvolutionAssistant } from '../agents/system/evolution-assistant.mjs'
+import {
+    mAssignEvolverListeners,
+    mChat,
+    mCreateAssistant,
+    mGetAssistant,
+    mGetChatCategory,
+    mMessages,
+    mRuns,
+ } from './class-avatar-functions.mjs'
 import { _ } from 'ajv'
 import { parse } from 'path'
 //  function definitions to extend remarkable classes
 function extendClass_avatar(_originClass,_references) {
     class Avatar extends _originClass {
-        #activeChatCategory = mGetMyLifeChatCategory()
+        #activeChatCategory = mGetChatCategory()
         #emitter = new EventEmitter()
         #evolver
         #factory
@@ -30,20 +38,13 @@ function extendClass_avatar(_originClass,_references) {
             return this
         }
         /* public functions */
-        // todo: move almost all logic to modular functionality to ease burden on class
-        async cancelRun(_run_id){	//	returns openai run object
-            return await this.openai.beta.threads.runs.cancel(
-                this.thread.id,
-                _run_id
-            )
-        }
         /**
          * Processes and executes incoming category set request.
          * @public
          * @param {string} _category - The category to set { category, contributionId, question }.
          */
         setActiveCategory(_category){
-            const _proposedCategory = mGetMyLifeChatCategory(_category)
+            const _proposedCategory = mGetChatCategory(_category)
             /* evolve contribution */
             if(_proposedCategory?.category){ // no category, no contribution
                 this.#evolver.setContribution(this.#activeChatCategory, _proposedCategory)
@@ -55,110 +56,18 @@ function extendClass_avatar(_originClass,_references) {
          * @param {object} ctx - The context object.
          * @returns {object} - The response(s) to the chat request.
         */
-        async chatRequest(ctx){ // todo: determine if ctx.state is sufficient
-            const _processStartTime = Date.now()
+        async chatRequest(ctx){
             if(!ctx?.state?.chatMessage)
                 throw new Error('No message provided in context')
             if(!this.thread)
                 this.thread = ctx.state.thread
-            //  add metadata, optional
-            //	assign uploaded files (optional) and push `retrieval` to tools
-            //	create message
-            const _chatMessage = ctx.state.chatMessage
-            this.setActiveCategory(_chatMessage) // also sets evolver contribution
-            const _message = new (this.factory.message)(_chatMessage)
-            await _message.init(this.thread) //  by embedding content above, init routine requires no parameters
-            this.messages.unshift(_message)
-            //	run thread
-            await this.run()
-            //	get message data from thread
-            const _messages = (await this.getMessages())
-                .filter(_msg => _msg.run_id == this.runs[0].id)
-                .map(_msg => {
-                    return new (this.factory.message)({
-                        avatar_id: this.id,
-                        content: _msg.content[0].text.value,
-                        mbr_id: this.mbr_id,
-                        role: 'assistant',
-                    })
-                })
-            const _responses = await Promise.all(
-                _messages.map(async _msg => {
-                    return await _msg.init(this.thread)
-                })
+            this.setActiveCategory(ctx.state.chatMessage) // also sets evolver contribution
+            return mChat(
+                this.#openai, 
+                this, 
+                ctx.state.chatMessage, 
+                ctx.state.MemberSession.conversation.id
             )
-            _responses.forEach(_msg => {
-                this.#evolver?.setContribution(this.#activeChatCategory, _msg)??false
-                this.messages.unshift(_msg)
-            })
-            //	update cosmos
-            if ((this?.factory !== undefined) && (process.env?.MYLIFE_DB_ALLOW_SAVE ?? false)) {
-                await this.factory.dataservices.patchArrayItems(
-                    ctx.state.MemberSession.conversation.id,
-                    'messages',
-                    [..._responses, _message]
-                )
-            }
-            //	return response
-            return _responses
-                .map(_msg=>{
-                    const __message = mPrepareMessage(_msg.text) // returns object { category, content }
-                    return {
-                        agent: 'server',
-                        category: __message.category,
-                        contributions: [],
-                        message: __message.content,
-                        response_time: Date.now()-_processStartTime,
-                        purpose: 'chat response',
-                        type: 'chat',
-                    }
-                })
-        }
-        async checkStatus(_thread_id,_run_id,_callInterval){
-            //	should be able to remove params aside from _callInterval, as they are properties of this
-            const _run = await this.openai.beta.threads.runs.retrieve(
-                _thread_id,
-                _run_id
-            )
-            switch(_run.status){
-                //	https://platform.openai.com/docs/assistants/how-it-works/runs-and-run-steps
-                case 'completed':
-                    if(!this?.runs[0]?.id === _run_id) this.runs.unshift(_run)	//	add
-                    else this.runs[0] = _run	//	update
-                    return true
-                case 'failed':
-                case 'cancelled':
-                case 'expired':
-                    return false
-                case 'queued':
-                case 'requires_action':
-                case 'in_progress':
-                case 'cancelling':
-                default:
-                    console.log(`...${_run.status}:${_thread_id}...`)
-                    break
-            }
-        }
-        async completeRun(_run_id){
-            return new Promise((resolve, reject) => {
-                const checkInterval = setInterval(async () => {
-                    try {
-                        const status = await this.checkStatus(this.thread.id, _run_id)
-                        if (status) {
-                            clearInterval(checkInterval)
-                            resolve('Run completed')
-                        }
-                    } catch (error) {
-                        clearInterval(checkInterval)
-                        reject(error)
-                    }
-                }, 700)
-                // Set a timeout to resolve the promise after 12 seconds
-                setTimeout(() => {
-                    clearInterval(checkInterval)
-                    resolve('Run completed (timeout)')
-                }, 220000)
-            })
         }
         /**
          * Proxy for emitter.
@@ -173,24 +82,7 @@ function extendClass_avatar(_originClass,_references) {
         async getAssistant(_dataservice){	//	openai `assistant` object
             //	3 states: 1) no assistant, 2) assistant id, 3) assistant object
             if(!this.assistant?.id.length) {
-                const _core = {
-                    name: this?.names[0]??this.name,
-                    model: process.env.OPENAI_MODEL_CORE,
-                    description: this.description,
-                    instructions: this.purpose,
-                    metadata: {
-                        ...Object.entries(this.categories)
-                            .filter(([key, value]) => this[value.replace(' ', '_').toLowerCase()]?.length)
-                            .slice(0, 16)
-                            .reduce((obj, [key, value]) => ({
-                                ...obj,
-                                [value]: this[value.replace(' ', '_').toLowerCase()],
-                            }), {})
-                    },
-                    file_ids: [],	//	no files at birth, can be added later
-                    tools: [],	//	only need tools if files
-                }
-                this.assistant = await this.openai.beta.assistants.create(_core)
+                this.assistant = await mCreateAssistant(this.#openai, this)
                 //	save id to cosmos
                 _dataservice.patch(this.id, {
                     assistant: {
@@ -198,58 +90,23 @@ function extendClass_avatar(_originClass,_references) {
                     ,	object: 'assistant'
                     }
                 })
+            } else if(!this.assistant?.name?.length){
+                this.assistant = await mGetAssistant(this.#openai, this.assistant.id)
             }
-            else if(!this.assistant?.name.length) this.assistant = await this.openai.beta.assistants.retrieve(this.assistant.id)
             return this.assistant
         }
         async getMessage(_msg_id){	//	returns openai `message` object
-            return this.messages.data.filter(_msg=>{ return _msg.id==_msg_id })
+            return this.messages.data
+                .filter(_msg=>{ return _msg.id==_msg_id })
         }
         async getMessages(){
-            this.messages = ( await this.openai.beta.threads.messages.list(
-                this.thread.id
-            ) )	//	extra parens to resolve promise
+            this.messages = ( await mMessages(this.#openai, this.thread.id) )
                 .data
+            console.log('getMessages', this.messages)
             return this.messages
-        }
-        async getRun(_run_id){	//	returns openai `run` object
-            return this.runs.filter(_run=>{ return _run.id==_run_id })
-        }
-        async getRuns(){	//	runs are also descending
-            if(!this.runs){
-                this.runs = await this.openai.beta.threads.runs.list(this.thread.id)
-                //	need to winnow to mapped array?
-            }
-        }
-        async getRunStep(_run_id,_step_id){
-            //	pull from known runs
-            return this.runs
-                .filter(_run=>{ return _run.id==_run_id })
-                .steps
-                    .filter(_step=>{ return _step.id==_step_id })
-        }
-        async getRunSteps(_run_id){
-            //	always get dynamically
-            const _run = this.runs.filter(_run=>{ return _run.id==_run_id })
-            _run.steps = await openai.beta.threads.runs.steps.list(this.thread.id, _run.id)
         }
         on(_eventName, listener){
             this.#emitter.on(_eventName, listener)
-        }
-        async run(){
-            const _run = await this.startRun()
-            if(!_run) throw new Error('Run failed to start')
-            this.runs = this?.runs??[]	//	once begun, ought complete even if failed
-            this.runs.unshift(_run)
-            // ping status
-            await this.completeRun(_run.id)
-        }
-        async startRun(){	//	returns openai `run` object
-            if(!this.thread || !this.messages.length) return
-            return await this.openai.beta.threads.runs.create(
-                this.thread.id,
-                { assistant_id: this.assistant.id }
-            )
         }
         /* getters/setters */
         get avatar(){
@@ -275,9 +132,9 @@ function extendClass_avatar(_originClass,_references) {
         /**
          * Set incoming contribution.
          * @param {object} _contribution
-         */
-        set contribution(_contribution){
-            this.#evolver.contribution = _contribution
+        */
+       set contribution(_contribution){
+           this.#evolver.contribution = _contribution
         }
         // todo: deprecate to available convenience public emit() function
         get emitter(){  //  allows parent to squeeze out an emitter
@@ -286,8 +143,12 @@ function extendClass_avatar(_originClass,_references) {
         get factory(){
             return this.#factory
         }
-        get openai(){
-            return this.#openai
+        /**
+         * Get uninstantiated class definition for message.
+         * @returns {class} - class definition for message
+         */
+        get message(){
+            return this.factory.message
         }
     }
     console.log('Avatar class extended')
@@ -324,7 +185,7 @@ function extendClass_contribution(_originClass,_references) {
          */
         async init(_factory){
             this.#factory = _factory
-            this.request.questions = await mGetQuestions(this, this.openai) // generate question(s) from cosmos or openAI
+            this.request.questions = await mGetQuestions(this, this.#openai) // generate question(s) from cosmos or openAI
             this.id = this.factory.newGuid
             this.status = 'prepared'
             this.emit('on-contribution-new',this)
@@ -437,8 +298,8 @@ function extendClass_message(_originClass,_references) {
         }
         /* public functions */
         async init(_thread){
-            if(!this.content & !this.files)
-                throw new Error('No content provided for message')
+            if((!this.content & !this.files)||!_thread)
+                throw new Error('Insufficient data provided for message init()')
             /* todo: limit requirements for context window, generate file and attach to file_ids array
             if(this.content.length > this.#maxContextWindow) {
                 //  TODO: generate file and attach to file_ids array
@@ -497,71 +358,6 @@ function extendClass_message(_originClass,_references) {
     return Message
 }
 /* modular functions */
-/* avatar modular functions */
-function mAssignEvolverListeners(_evolver, _avatar){
-    /* assign evolver listeners */
-    _evolver.on(
-        'on-contribution-new',
-        _contribution=>{
-            _contribution.emit('on-contribution-new', _contribution)
-        }
-    )
-    _evolver.on(
-        'avatar-change-category',
-        (_current, _proposed)=>{
-            _avatar.category = _proposed
-            console.log('avatar-change-category', _avatar.category.category)
-        }
-    )
-    _evolver.on(
-        'on-contribution-submitted',
-        _contribution=>{
-            // send to gpt for summary
-            const _responses = _contribution.responses.join('\n')
-            const _summary = _avatar.factory.openai.completions.create({
-                model: 'gpt-3.5-turbo-instruct',
-                prompt: 'summarize answers in 512 chars or less, if unsummarizable, return "NONE": ' + _responses,
-                temperature: 1,
-                max_tokens: 700,
-                frequency_penalty: 0.87,
-                presence_penalty: 0.54,
-            })
-            // evaluate summary
-            console.log('on-contribution-submitted', _summary)
-            abort
-            //  if summary is good, submit to cosmos
-        }
-    )
-}
-function mFormatCategory(_category){
-    return _category
-        .trim()
-        .slice(0, 128)  //  hard cap at 128 chars
-        .replace(/\s+/g, '_')
-        .toLowerCase()
-}
-/**
- * Returns chat category object
- * @modular
- * @param {object} _category - local category { category, contributionId, question }
- * @returns {object} - local category { category, contributionId, messages }
- */
-function mGetMyLifeChatCategory(_category) {
-    const _proposedCategory = {
-        category: '',
-        contributionId: undefined,
-        content: undefined,
-    }
-    if(_category?.category && _category.category.toLowerCase() !== 'off'){
-        _proposedCategory.category = mFormatCategory(_category.category)
-        _proposedCategory.contributionId = _category.contributionId
-        _proposedCategory.content = 
-            _category?.question??
-            _category?.message??
-            _category?.content // test for undefined
-    }
-    return _proposedCategory
-}
 /* contribution modular functions */
 /**
  * Evaluates Contribution and may update `status` property.
@@ -707,31 +503,6 @@ function mReviewContent(_message, _obj){
             _obj?.content??
             _message?.content??
             ''
-}
-/**
- * returns simple micro-message with category after logic mutation. Currently tuned for openAI gpt-assistant responses.
- * @modular
- * @param {string} _msg text of message, currently from gpt-assistants
- * @returns {object} { category, content }
- */
-function mPrepareMessage(_msg){
-    /* parse message */
-    // Regular expression to match the pattern "Category Mode: {category}. " or "Category Mode: {category}\n"; The 'i' flag makes the match case-insensitive
-    const _categoryRegex = /^Category Mode: (.*?)\.?$/gim
-    const _match = _categoryRegex.exec(_msg)
-    const _messageCategory = mFormatCategory(_match?.[1]??'')
-    // Remove from _msg
-    _msg = _msg.replace(_categoryRegex, '')
-    const _content = _msg.split('\n')
-        .filter(_line => _line.trim() !== '') // Remove empty lines
-        .map(_msg=>{
-            return new Marked().parse(_msg)
-        })
-        .join('\n')
-    return {
-        category: _messageCategory,
-        content: _content,
-    }
 }
 /* exports */
 export {
